requirements.deploy.txt
# === API / Framework ===
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1

# === LLM / Embedding / NLP ===
openai==1.67.0
chromadb==0.5.20
sentence-transformers==3.2.0
scikit-learn==1.7.2
numpy==2.1.1
keybert==0.8.5
kiwipiepy==0.19.0
kiwipiepy_model==0.19.0
k-means-constrained==0.7.6

# === LangChain ===
langchain==0.3.21
langchain-core==0.3.46
langchain-openai==0.3.9
langchain-text-splitters==0.3.7

# === Utility ===
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.3
tqdm==4.67.1
APScheduler==3.10.4
